refactor(scripts): _compute_derived_status verhuist naar track_reconciler_status.py
Fase 1 PR 1 van track file-size-refactor-debt (plan: claudedocs/REFACTOR_PLAN.md
par.4.1). Pure verplaatsing, geen regel logica gewijzigd:

- Nieuwe platte module scripts/lib/track_reconciler_status.py (176 regels) met
  de functie byte-identiek (AST-bewezen via refactor_equivalence.py).
- track_reconciler.py re-exporteert de naam onderaan de module zodat geen
  consumer hoeft te wijzigen (refactor_surface.py: zelfde object, exit 0).
- Late binding in beide richtingen doorbreekt de import-cyclus: de nieuwe
  module importeert zijn helpers (_has_col, _parse_pr_numbers, de
  dispatch-state constanten) na de functiedefinitie.
- Size-gate: 1143 -> 1006 regels, blijft warning (drempel 500/1200), geen
  allowlist nodig.

Bewijslast (zes bewijzen) in het unified report van dispatch
20260802-f1pr1-track-reconciler-status.

Dispatch-ID: 20260802-f1pr1-track-reconciler-status
